Added support for Glyphicons and Font Awesome icon solutions
This adds a new process to grab the before: pseudo element for only objects with a font awesome or glyphicon class and inject that content into the copied div that is destined for print. In the future, this should be expanded to inject both before and after pseudo elements and retain the object they are being specified on. For this commit, this is only for empty objects (span for Glyphicon, i for Font Awesome) in which the class is specified as in those respective libraries.